Updated models and services with correct data types (#28)
* Updated invoice model properties to correct data types

* Updated order model properties with correct data types

* Added CustomJsonConverter to InvoiceRow

* Updated CreateInvoiceTest

* Replaced float in favor for decimal in models (better precision)

* Fixed build error (decimal/double) in test project (GetSupplierInvoicePaymentTest)

+++ Fortnox.NET/Models/AttendanceTransaction/AttendanceTransactionSubset.cs
@@ -17,7 +17,7 @@ namespace FortnoxNET.Models.AttendanceTransaction
 		public string EmployeeId { get; set; }
 		public string CauseCode { get; set; }
 		public DateTime Date { get; set; }
-		public float Hours { get; set; }
+		public decimal Hours { get; set; }
 		public string CostCenter { get; set; }
 		public string Project { get; set; }
 	}
