+++ Fortnox.NET/Models/Price/Price.cs
@@ -16,13 +16,13 @@ namespace FortnoxNET.Models.Price
         [JsonReadOnly]
         public DateTime? Date { get; set; }
 
-        public float FromQuantity { get; set; }
+        public decimal FromQuantity { get; set; }
 
-        public float Percent { get; set; }
+        public decimal Percent { get; set; }
 
         public string PriceList { get; set; }
 
         [JsonProperty(PropertyName = "Price")]
-        public float PriceValue { get; set; }
+        public decimal PriceValue { get; set; }
     }
 }
